Gates crated

+++ app/Http/Controllers/Api/AttendeeController.php
@@ -8,12 +8,18 @@ use App\Http\Traits\CanLoadRelationships;
 use App\Models\Attendee;
 use App\Models\Event;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class AttendeeController extends Controller
 {
     use CanLoadRelationships;
 
     private array $relations = ['user'];
+
+    public function __construct()
+    {
+        $this->middleware('auth:sanctum')->except(['index', 'show', 'update']);
+    }
 
     public function index(Event $event)
     {
@@ -44,8 +50,11 @@ class AttendeeController extends Controller
         );
     }
 
-    public function destroy(string $event, Attendee $attendee)
+    public function destroy(Event $event, Attendee $attendee)
     {
+        if (Gate::denies('delete-attendee', [$event, $attendee])) {
+            abort(403, 'you are not allowed to delete attendee');
+        }
         $attendee->delete();
 
         return response()->json([
